*Description of changes:*

In #790 (released in 1.11.5) we made `IonWriter.writeValues(IonReader)` iterative instead of recursion in order to eliminate unbounded recursion from the library.

While working on #1093 I noticed that this had caused negative performance impact for certain datasets:

Dataset: deeply nested single value

Benchmarking code: [ion-java-benchmark-cli](https://github.com/amazon-ion/ion-java-benchmark-cli/tree/read-tests-writevalues), modified so that the `read` command tests `IonWriter.writeValues(IonReader)` with stream copy optimization *disabled*. I will formally add an option to the tool that does this in a separate PR.

CLI command: `./benchmark-cli read --mode AverageTime --time-unit microseconds --iterations 2 --warmups 2 --forks 2 --ion-reader non_incremental  --io-type buffer <file>`

1.11.4
Time: 38.788 us/op

1.11.5
Time: 42.918 us/op

1.11.8 (another regression happened here, related to the change to how PatchPoints are handled. This will be addressed separately)
Time: 44.937 us/op

This PR:
Time: 40.474 us/op

I identified this change by comparing CPU profiles. In 1.11.5, `IonWriter.setFieldNameSymbol` showed up at the top of the profile, whereas in 1.11.4 it was much lower. This is because in 1.11.4 it was JIT compiled, whereas in 1.11.5 it was simply inlined, which is much slower. JIT compilation can be limited by the size of the method, so I tried this and it worked.


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.